+++ chain/src/indexer/durable/mod.rs
@@ -1,0 +1,5 @@
+mod drainer;
+mod state;
+
+pub(crate) use drainer::Drainer;
+pub(crate) use state::{DrainerMetrics, Enqueuer, FinalizedEntry, SharedUploadState, UploadState};
